ConnectButtonRenderer.isConnecting via useAccount
The previous commit had incorrectly used useConnect to expand coverage for isConnecting.

useConnect is incorrect because its isLoading true if and only if that instance of useConnect is
connecting. I.e. useConnect isLoading is a local and not global status.

Instead, we use useAccount.isConnecting/isReconnecting, which are correctly global statuses.