feat: add Persian localization support
Persian (fa-IR) language option is added to the application and all relevant messages have been provided for this supported language. This will ensure a larger user base and allow Persian speaking users to interact with the application in their native language. The localization files for Persian have also been added to stay consistent with the internationalization process.